+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Allow defining baseTags property in addons. Will be extended with "tags" option when instantiated for calculating provider tags. (#143)
 - feat: Add tags getter
 - feat: Add "selector" base tag to selectors (#163)
+- feat: Add `cleanDependenciesCache` method to `providers` handler (#164)
 ### Changed
 ### Fixed
 - docs(readme): Fix typos