+++ CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- feat: Allow defining baseTags property in addons. Will be extended with "tags" option when instantiated for calculating provider tags. (#143)
+- feat: Add tags getter
+- feat: Add "selector" base tag to selectors (#163)
+- feat: Add `cleanDependenciesCache` method to `providers` handler (#164)
 ### Changed
 ### Fixed
+- docs(readme): Fix typos
 ### Removed
 
 ## [2.8.2] - 2020-12-12